fixed file names.  fixed shift_array code. ran tests.
