--- challenges/binary_search/README.md
@@ -1,13 +1,0 @@
-# Data Structure & Algorithms
-A repository dedicated to my passion and appreciation for all things data structures and algorithms!
-
-## Installation
-Follows a standard GitHub workflow for cloning and requiring implementations using my favorite language of choice, wat.
-
-# Binary Search
-
-## Challenge
-Write a function called BinarySearch which takes in 2 parameters: a sorted array and the search key. Without utilizing any of the built-in methods available to your language, return the index of the array’s element that is equal to the search key, or -1 if the element does not exist.
-
-## Solution
-data-structures-and-algorithms/assets/binary_search.jpg
