+++ challenges/largest-product_array/README.md
@@ -1,0 +1,12 @@
+# RETURN LARGEST PRODUCT
+
+## Challenge
+Write a function called LargestProduct which takes in a 2D array. Without utilizing any of the built-in methods available to your language, return the largest product of 2 adjacent values within the 2D array.
+## Example data:
+```
+Input	                                       Output
+[ [ 1, 2 ], [ 3, 4 ], [ 5, 6 ], [ 7, 8 ] ];    56
+```
+
+## Solution
+data-structures-and-algorithms/assets/largest_product_array.jpg
