--- challenges/shift_array/README.md
@@ -1,15 +1,0 @@
-# Data Structure & Algorithms
-A repository dedicated to my passion and appreciation for all things data structures and algorithms!
-
-## Installation
-Follows a standard GitHub workflow for cloning and requiring implementations using my favorite language of choice, wat.
-
-
-# Insert and shift middle index of array
-take an array and addtional integer and insert the integer into the middle of the array.
-
-## Challenge
-Write a function called insertShiftArray that does not use packages to solve for the problem. 
-
-## Solution
-./data-structures-and-algorithms/assets/shift-array.jpg
